refactor: Organize Category domain
