+++ frontend/src/features/category/types/Category.ts
