+++ frontend/src/features/category/types/CategoryHeaderProps.ts
