+++ frontend/src/features/category/types/CategoryItemProps.ts
