+++ frontend/src/features/category/types/CategorySearchProps.ts
